backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0

# WebSocket
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-socketio[client]==5.11.0
gevent==24.2.1
gevent-websocket==0.10.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Redis & Caching
redis==5.0.1
Flask-Caching==2.1.0

# Celery (Async Tasks)
celery==5.3.4
celery[redis]==5.3.4

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7

# AWS SDK
boto3==1.34.11
botocore==1.34.11

# Supabase
supabase==2.3.0
postgrest-py==0.13.0

# HTTP & API
requests==2.31.0
httpx==0.26.0

# WebRTC & Agora
agora-token-builder==1.0.0

# AI & ML
openai==1.7.0  # For Grok API compatibility

# Data Validation
marshmallow==3.20.1
python-dotenv==1.0.0
pydantic==2.5.3

# Utils
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0
uuid==1.30

# Production Server
gunicorn==21.2.0
eventlet==0.35.1

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.2
